Refactor SalesOrder filters
